LPS-142946 Move *Helper classes to '.helper'
