+++ modules/apps/portal-search/portal-search-web/src/main/java/com/liferay/portal/search/web/internal/low/level/search/options/portlet/preferences/LowLevelSearchOptionsPortletPreferencesImpl.java
@@ -22,8 +22,8 @@ import com.liferay.portal.kernel.json.JSONUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.search.web.internal.helper.PortletPreferencesHelper;
 import com.liferay.portal.search.web.internal.search.options.portlet.SearchOptionsPortletPreferences;
-import com.liferay.portal.search.web.internal.util.PortletPreferencesHelper;
 
 import java.util.Optional;
