+++ modules/apps/portal-search/portal-search-web/src/main/java/com/liferay/portal/search/web/internal/search/options/portlet/SearchOptionsPortletPreferencesImpl.java
@@ -15,7 +15,7 @@
 package com.liferay.portal.search.web.internal.search.options.portlet;
 
 import com.liferay.petra.string.StringPool;
-import com.liferay.portal.search.web.internal.util.PortletPreferencesHelper;
+import com.liferay.portal.search.web.internal.helper.PortletPreferencesHelper;
 
 import java.util.Optional;
