+++ modules/apps/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/indexer/helper/PreFilterContributorHelper.java
@@ -12,31 +12,27 @@
  * details.
  */
 
-package com.liferay.portal.search.internal.searcher;
+package com.liferay.portal.search.internal.indexer.helper;
 
-import com.liferay.portal.kernel.search.Hits;
-import com.liferay.portal.kernel.search.Query;
+import com.liferay.portal.kernel.search.Indexer;
 import com.liferay.portal.kernel.search.SearchContext;
+import com.liferay.portal.kernel.search.filter.BooleanFilter;
+import com.liferay.portal.search.spi.model.registrar.ModelSearchSettings;
 
-import java.util.List;
 import java.util.Map;
 
 /**
  * @author André de Oliveira
  */
-public interface IndexSearcherHelper {
+public interface PreFilterContributorHelper {
 
-	public String getQueryString(SearchContext searchContext, Query query);
+	public void contribute(
+		BooleanFilter booleanFilter,
+		Map<String, Indexer<?>> entryClassNameIndexerMap,
+		SearchContext searchContext);
 
-	public Hits search(SearchContext searchContext, Query query);
-
-	public long searchCount(SearchContext searchContext, Query query);
-
-	public String spellCheckKeywords(SearchContext searchContext);
-
-	public Map<String, List<String>> spellCheckKeywords(
-		SearchContext searchContext, int max);
-
-	public String[] suggestKeywordQueries(SearchContext searchContext, int max);
+	public void contribute(
+		BooleanFilter booleanFilter, ModelSearchSettings modelSearchSettings,
+		SearchContext searchContext);
 
 }
